Improve test CMLs and add missing tests (RttrConfig and System)
