+++ tests/CMakeLists.txt
@@ -1,9 +1,12 @@
+include(AddTestcase)
+
 add_subdirectory(common)
 add_subdirectory(languages)
 add_subdirectory(legacyFiles)
 add_subdirectory(libGameData)
 add_subdirectory(libsamplerate_cpp)
 add_subdirectory(mockupDrivers)
+add_subdirectory(rttrConfig)
 add_subdirectory(s25client)
 add_subdirectory(s25Main)
 add_subdirectory(testHelpers)
