+++ tests/common/CMakeLists.txt
@@ -1,4 +1,3 @@
-include(AddTestcase)
 add_testcase(NAME common
     LIBS s25Common testHelpers
 )
