+++ tests/libGameData/CMakeLists.txt
@@ -1,4 +1,3 @@
-include(AddTestcase)
 add_testcase(NAME gameData
     LIBS gamedata legacyTestfiles testHelpers
 )
