+++ tests/libsamplerate_cpp/CMakeLists.txt
@@ -1,4 +1,3 @@
-include(AddTestcase)
 add_testcase(NAME samplerate_cpp
     LIBS samplerate_cpp
 )
