+++ tests/rttrConfig/CMakeLists.txt
@@ -1,4 +1,3 @@
-include(AddTestcase)
 add_testcase(NAME rttrConfig
-    LIBS s25Config
+    LIBS rttrConfig testHelpers
 )
