+++ tests/s25Main/CMakeLists.txt
@@ -25,8 +25,6 @@ target_link_libraries(testWorldFixtures PUBLIC s25Main PRIVATE testHelpers)
 target_include_directories(testWorldFixtures PUBLIC .)
 enable_warnings(testWorldFixtures)
 
-include(AddTestcase)
-
 add_subdirectory(audio)
 add_subdirectory(drivers)
 add_subdirectory(integration)
